requirements.txt
django-oscar
pycountry
